# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Trainerplan API
  version: 1.0.0
  description: |
    Versionierte HTTP-API für die Trainerplan-Web-App und native Clients.
    Browser verwenden eine HTTP-only-Cookie-Session; native Clients verwenden
    das beim Login ausgegebene Bearer-Token.
servers:
  - url: /api/v1
    description: Aktuelle API-Version
...
security:
  - bearerAuth: []
  - sessionCookie: []
paths:
...
  /training-plans:
    get:
      tags: [Trainingspläne]
      summary: Trainingspläne laden
      responses:
        '200':
          description: Trainingspläne
          content:
            application/json: { schema: { $ref: '#/components/schemas/PlansPayload' } }
    put:
      tags: [Trainingspläne]
      summary: Trainingspläne speichern
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/PlansPayload' } }
      responses:
        '200': { description: Trainingspläne gespeichert, content: { application/json: { schema: { $ref: '#/components/schemas/PlansPayload' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /users:
    get:
      tags: [Mannschaft]
      summary: Sichtbare Benutzer laden
      responses: { '200': { description: Benutzer geladen, content: { application/json: { schema: { type: object, properties: { users: { type: array, items: { $ref: '#/components/schemas/User' } } } } } } } }
    put:
      tags: [Mannschaft]
      summary: Benutzerprofile speichern
      requestBody:
        required: true
        content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/User' } } } }
      responses: { '200': { description: Benutzer gespeichert, content: { application/json: { schema: { type: object, properties: { users: { type: array, items: { $ref: '#/components/schemas/User' } } } } } } } }
...
  /templates:
    get:
      tags: [Vorlagen]
      summary: Trainingsvorlagen laden
      responses: { '200': { description: Vorlagen geladen, content: { application/json: { schema: { type: object, properties: { templates: { type: array, items: { $ref: '#/components/schemas/Template' } } } } } } } }
    post:
      tags: [Vorlagen]
      summary: Trainingsvorlage erstellen
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Template' } } }
      responses: { '201': { description: Vorlage erstellt, content: { application/json: { schema: { type: object, properties: { template: { $ref: '#/components/schemas/Template' } } } } } } }
  /templates/{id}:
    parameters: [{ $ref: '#/components/parameters/ResourceId' }]
    patch:
      tags: [Vorlagen]
      summary: Trainingsvorlage ändern
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Template' } } }
      responses: { '200': { description: Vorlage geändert, content: { application/json: { schema: { type: object, properties: { template: { $ref: '#/components/schemas/Template' } } } } } } }
    delete:
      tags: [Vorlagen]
      summary: Trainingsvorlage löschen
      responses: { '200': { $ref: '#/components/responses/Ok' } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: session-token }
    sessionCookie: { type: apiKey, in: cookie, name: trainerplan_session }
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 160 }
  responses:
    Ok:
      description: Erfolgreich
      content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } }
    BadRequest:
      description: Ungültige Anfrage
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    Forbidden:
      description: Keine Berechtigung
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties: { error: { type: string } }
    Ok:
      type: object
      properties: { ok: { type: boolean, example: true } }
    Role: { type: string, enum: [admin, trainer, player] }
...
    User:
      type: object
      required: [id, name, email, role, position, phone, birthday, ageGroup, dribblingRating, shootingRating, passingRating]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/Role' }
        position: { type: string }
        number: { type: [integer, 'null'] }
        ballNumber: { type: [integer, 'null'] }
        phone: { type: string }
        birthday: { type: string, format: date }
        ageGroup: { type: string }
        avatar: { type: [string, 'null'] }
        groupId: { type: [string, 'null'] }
        dribblingRating: { type: integer, minimum: 0, maximum: 5 }
        shootingRating: { type: integer, minimum: 0, maximum: 5 }
        passingRating: { type: integer, minimum: 0, maximum: 5 }
        internalTeam: { type: [string, 'null'], enum: [A, B, null] }
...
    Exercise:
      type: object
      required: [id, title, description, duration, players, ageGroup, ageRange, category, accent, intensity, focus, setup, coaching, materials, fieldSize, variant]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        duration: { type: integer, minimum: 1, maximum: 300 }
        players: { type: string }
        ageGroup: { type: string, enum: [F-Jugend] }
        ageRange: { type: string, enum: [U8, U9, U8/U9] }
        category: { type: string, enum: [Ankommen, Einstieg, Hauptteil, Abschlussspiel] }
        accent: { type: string, pattern: '^#[0-9a-fA-F]{6}$' }
        intensity: { type: string, enum: [Niedrig, Mittel, Hoch] }
        focus: { type: array, items: { type: string } }
        setup: { type: string }
        coaching: { type: array, items: { type: string } }
        materials: { type: array, items: { type: object, required: [id, count], properties: { id: { type: string }, count: { type: integer, minimum: 1 } } } }
        fieldSize: { type: string }
        variant: { type: integer, minimum: 0 }
        youtubeUrl: { type: string, format: uri }
        trainerId: { type: [string, 'null'] }
        internalTeam: { type: [string, 'null'], enum: [A, B, null] }
    Plans:
      type: object
      additionalProperties:
        type: array
        items: { $ref: '#/components/schemas/Exercise' }
    PlanMeta:
      type: object
      required: [name, focus]
      properties: { name: { type: string }, focus: { type: array, items: { type: string } } }
    PlansPayload:
      type: object
      required: [plans, planMeta]
      properties:
        plans: { $ref: '#/components/schemas/Plans' }
        planMeta: { type: object, additionalProperties: { $ref: '#/components/schemas/PlanMeta' } }
...
    Template:
      type: object
      required: [id, name, kind, exercises]
      properties:
        id: { type: string }
        name: { type: string }
        kind: { type: string, enum: [plan, phase] }
        phase: { type: string }
        focus: { type: array, items: { type: string } }
        autoApply: { type: boolean }
        exercises: { type: array, items: { $ref: '#/components/schemas/Exercise' } }
...